refactor: split up keyboard navigation
